Add Annotations support.
Update the conversation monitor with the latest annotations support in Scrible.
Annotations plugins implemented as an governance interceptors to handle annotations of type:
{assertion:....} and {commitment: ...}
AssertionInterceptor evaluates a python expression on the message payload
CommitmentInterceptor is added to teh codebase, but is not implemented yet.

Note: To use any of the annotation plugins you should register the respective interceptors
in your local pyon.yml file.